feat: enhance meal management functionality and update token models

+++ backend/api/api.py
@@ -1,12 +1,12 @@
 from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
-from sqlalchemy import select
-from starlette.status import HTTP_201_CREATED, HTTP_406_NOT_ACCEPTABLE
+from sqlalchemy import select, update
+from fastapi import status
+
 
 from backend.api.login import get_current_user
 from backend.models.tokens import JWTAccessBase
 from backend.schemas.meals_eaten import MealsEaten
-from backend.schemas.user import User
 
 from ..schemas import MohMitzrachim
 from ..models.food import FoodDetail, MealEntry
@@ -14,6 +14,9 @@ from ..db import session
 
 
 router = APIRouter()
+
+# FIXME: Don't use / , use a query param.
+# example. q={QUERY}
 
 
 @router.get('/foods/{food_query}', response_model_by_alias=False, response_model=dict[str, list[FoodDetail]])
@@ -23,14 +26,14 @@ def query_foods(food_query: str, current_user: Annotated[JWTAccessBase, Depends(
     raise HTTPException(404, "food type wasnt found")
 
 
-@router.put('/meal', response_model_by_alias=False,  status_code=HTTP_201_CREATED)
+@router.put('/meal', response_model_by_alias=False,  status_code=status.HTTP_201_CREATED)
 def add_meal(
     current_user: Annotated[JWTAccessBase, Depends(get_current_user)],
     meal: MealEntry
 ):
-    user_id = session.execute(select(User.id).where(
-        User.username == current_user.sub)).scalar_one()
-    db_meal: MealsEaten = MealsEaten(user_id=user_id,
+    """Creates a new meal data and sends it into the database"""
+
+    db_meal: MealsEaten = MealsEaten(user_id=current_user.sub,
                                      code_id=meal.food_id,
                                      mida_id=meal.mida_id,
                                      amount=meal.amount,
@@ -42,29 +45,46 @@ def add_meal(
     session.commit()
 
 
-@router.delete('/meal', response_model_by_alias=False,  status_code=HTTP_201_CREATED)
+@router.delete('/meal', response_model_by_alias=False,  status_code=status.HTTP_201_CREATED)
 def delete_meal(
     current_user: Annotated[JWTAccessBase, Depends(get_current_user)],
     meal_id: int
 ):
+    """given a meal id, deletes a user meal"""
     # Check That it is indeed the corect user asking to delete.
-
-    #
     stmt = (
         select(MealsEaten).where(MealsEaten.id == meal_id)
     )
 
     db_meal = session.execute(stmt).scalar_one_or_none()
 
-    if not db_meal:
-        raise HTTPException(HTTP_406_NOT_ACCEPTABLE, "Meal item doesn't exist")
+    if not db_meal or db_meal.user_id != current_user.sub:
+        raise HTTPException(status.HTTP_403_FORBIDDEN,
+                            "Meal item doesn't exist")
 
     session.delete(db_meal)
 
     session.commit()
 
-    #
 
-    # @router.put
-    # @router.delete
-    # @router.patch
+@router.patch('/meal', response_model_by_alias=False,  status_code=status.HTTP_200_OK)
+def update_meal(
+    current_user: Annotated[JWTAccessBase, Depends(get_current_user)],
+    meal: MealEntry,
+    meal_id: int
+):
+    """Update a meal value from a specific user"""
+    # 1. Check that mida_id Meals Eaten value is the correct user_id
+
+    db_meal = session.get(MealsEaten, meal_id)
+
+    if not db_meal or db_meal.user_id != current_user.sub:
+        raise HTTPException(status.HTTP_403_FORBIDDEN,
+                            "Meal item doesn't exist")
+
+    update_data = meal.model_dump(exclude_unset=True)
+
+    for k, v in update_data.items():
+        setattr(db_meal, k, v)
+
+    session.commit()
